use anonymous, not named, type constraints for one-offs
By their nature, named type constraints are global.  This means that it's
particularly important to choose names not likely to conflcit with any other
type name someone defined somewhere else (and is one of the primary reasons
for the existance of MooseX::Types).

For our purposes, it's enough to create anonymous types, that are then
referenced by 'constants' used during attribute definition.  This gives us the
type constraint/coercions we need without the risk of a potential (painful and
confusing debug session somewhere down the line :)