remove spaces from configs in metadata
A YAML line like

    "git --version": 1.2.3

is generated in newer versions of the YAML encoder, but these quotes are
unexpected in older versions of the decoder (CPAN::Meta::YAML 0.10 ->
0.11, i.e. YAML::Tiny 1.55 -> 1.60, but bisects to commit
73ea4027, i.e. YAML::Tiny 1.57 containing the changed behaviour).
Consequently Test::CPAN::Meta fails in older perls when checking a
META.yml that was generated in a modern perl.  This *probably* hasn't
actually caused problems in the wild, since the toolchain knows how to
fall back to makefile args encoded in the Makefile directly (i.e. to get
configure_requires prerequisites), but it's still best to just avoid the
situation.