test util: ensure that branch is called 'master'
`git init` determines the name of the initial branch based on the
contents of .git/HEAD, which is copied in from a template directory,
which can be set in a number of ways. I happen to use a non-standard
layout, so all of these tests failed for me (because they assume a
'master' branch). This commit just fixes up the initial branch name to
be what we expect, regardless of any local template directory.