Add option stage to Git::Init
At the moment, staging files happens automatically.
We want to allow user to control that.
By default the new functionality will retain the old behaviour.
Only by setting the new option to false, will the existing
behaviour change.

if true (the default), stage (add to Git index)
all files in the newly-minted dist directory. If commit option
is true and stage option is false, Git will create an empty commit.
An empty commit is a valid commit and can be used
as the first commit (root commit) in a repository.

Signed-off-by: Mikko Koivunalho <mikkoi@cpan.org>